extracted a class into separate file
